Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float x;
-    scanf("%.4f", &x);
-    printf("You entered: %.4f", x);
+    double x;
+    scanf("%f", &x);
+    printf("You entered: %f", x);
     return 0;
 }
